account for struct size/alignment under MSVC
GCC and Clang have some very weird behavior described in
https://github.com/rust-lang/rust-bindgen/issues/1753, which allows
aligned typedefs to have a size that's not a multiple of their
alignment. That leads to struct members "overlapping", which requires a
lot of manual tweaking in our bindings. Unfortunately, MSVC does *not*
display the same behavior, and we need more manual tweaking to handle
the difference there.